pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.8"
dependencies = [
    "autogen-agentchat[chromadb]>=0.4.9.2",
    "autogen-ext[mcp,openai,web-surfer]>=0.4.9.2",
    "chromadb>=0.6.3",
    "docker>=7.1.0",
    "ipykernel>=6.29.5",
    "ipynb>=0.5.1",
    "jupyter>=1.1.1",
    "jupyter-kernel-gateway>=3.0.1",
    "litestar[standard]>=2.15.1",
    "pip>=25.0.1",
    "python-dotenv>=1.0.1",
]
